requirements.txt
streamlit
python-dotenv
langchain
openai
langchain-core
langchain-community
langchain-openai
langchain-chroma
langchain-text-splitters
unstructured[all-docs]
langchainhub
python-magic-bin
